requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
youtube-transcript-api>=1.0.3
yt-dlp>=2025.3.12
ffmpeg-python>=0.2.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9
Pillow>=10.0.0
opencv-python>=4.10.0.84
numpy>=2.1.0
pytesseract>=0.3.10
faster-whisper>=1.0.0
requests>=2.32.0
pytest>=9.0.1
